Moved minigames to pages
