+++ frontend/src/pages/hotwire/board.css
