+++ frontend/src/pages/hotwire/hotwire_puzzles.json
